moved functions into one file
