+++ app/functions/index.js
@@ -24,9 +24,8 @@ const logger = require("firebase-functions/logger");
 const { getLinkedAccounts } = require('./frontendFunctions');
 const { createLinkToken, exchangePublicToken, loadAllUserTransactions } = require('./plaidFunctions');
 const { createPaymentIntent } = require('./stripeFunctions');
-const { triggerDailyLogs} = require('./transferDailyToHolding');
-const { triggerMonthlyLogs } = require('./generateMonthlyLogs');
-const { processMonthlyLog } = require('./transferToRecipients');
+const { triggerDailyLogs, triggerMonthlyLogs, processMonthlyLog} = require('./recipientMoneyLogAndMove');
+
 
 // Export the frontend functions
 exports.getLinkedAccounts = getLinkedAccounts;
@@ -39,11 +38,7 @@ exports.loadAllUserTransactions = loadAllUserTransactions;
 // Export the stripe functions
 exports.createPaymentIntent = createPaymentIntent;
 
-// Export the Daily Totals to Holding Account but not processed
+// Export the Daily and Monthly Totals to Holding Account but not processed but also a function to process the monthly logs
 exports.triggerDailyLogs = triggerDailyLogs
-
-// Export the Monthly Totals to Holding Account but not processed
 exports.triggerMonthlyLogs = triggerMonthlyLogs
-
-// Export the Monthly Totals to Holding Account and proccessed it to the recipients
 exports.processMonthlyLog = processMonthlyLog
