+++ app/package.json
@@ -15,6 +15,7 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@stripe/react-stripe-js": "^2.7.3",
     "@stripe/stripe-js": "^4.1.0",
+    "date-fns": "^3.6.0",
     "firebase": "^10.12.4",
     "firebase-admin": "^12.3.0",
     "plaid": "^26.0.0",
